refactor: extract reusable components from tab screens
Move FeaturedEventCard, BoardPeekRow, MiniEventRow from index.tsx into
components/home/. Extract EmptyPanel, SignInCallout, CreatePostSheet, and
shared ColorSet/GroupKind types from feed.tsx and chat.tsx into
components/connect/. Screen files reduced by ~900 lines total.
