+++ packages/frontend/components/connect/types.ts
@@ -1,0 +1,19 @@
+export type GroupKind = 'center' | 'event'
+
+export type ColorSet = {
+  page: string
+  surface: string
+  panel: string
+  rail: string
+  card: string
+  cardActive: string
+  border: string
+  borderStrong: string
+  text: string
+  textMuted: string
+  textSoft: string
+  orange: string
+  orangeSoft: string
+  green: string
+  greenSoft: string
+}
